+++ docs/prerender.js
@@ -1,47 +1,72 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const fs = require('fs')
 const path = require('path')
+const fg = require('fast-glob')
+const execa = require('execa')
+const { SitemapStream, streamToPromise } = require('sitemap')
 
 const toAbsolute = (p) => path.resolve(__dirname, p).replace(/\\/, '/')
 
-const manifest = require('./dist/static/ssr-manifest.json')
+const manifest = require(toAbsolute('dist/static/ssr-manifest.json'))
 const template = fs.readFileSync(toAbsolute('dist/static/index.html'), 'utf-8')
-const { render } = require('./dist/server/entry-server.js')
+const { render } = require(toAbsolute('dist/server/entry-server.js'))
+const files = fg.sync('docs/docs/**/!(ColumnSetting|Crud|TreeSelect).md')
 
 const writeFileRecursive = function (path, buffer) {
   const lastPath = path.substring(0, path.lastIndexOf('/'))
 
   fs.mkdir(lastPath, { recursive: true }, () => {
     fs.writeFileSync(path, buffer)
+    console.log('pre-rendered:', path)
   })
 }
 
-const fileDisplay = (file) => {
-  fs.readdirSync(toAbsolute(file)).forEach(async (filename) => {
-    const filedir = path.join(file, filename).replace(/\\/, '/')
+const getUpdatedTime = async (filePath) => {
+  const { stdout } = await execa('git', [
+    '--no-pager',
+    'log',
+    '-1',
+    '--format=%at',
+    filePath,
+  ])
+  const time = Number.parseInt(stdout, 10) * 1000
 
-    if (fs.statSync(toAbsolute(filedir)).isDirectory()) {
-      fileDisplay(filedir)
-    } else {
-      const url = filedir
-        .replace(/^docs/, '')
-        .replace(/\.(vue|md)$/, '')
-        .replace(/index$/, '')
-        .replace(/\/([^/]*)$/, (item) =>
-          item.replace(/\B([A-Z])/g, '-$1').toLowerCase()
-        )
-      const [appHtml, preloadLinks] = await render(url, manifest)
-      const html = template
-        .replace('<!--preload-links-->', preloadLinks)
-        .replace('<!--app-html-->', appHtml)
-
-      const filePath = `dist/static${url.replace(/\/$/, '/index')}.html`
-      writeFileRecursive(toAbsolute(filePath), html)
-      console.log('pre-rendered:', filePath)
-    }
-  })
+  return new Date(time).toISOString()
 }
 
-fileDisplay('docs')
+;(async () => {
+  const stream = new SitemapStream({ hostname: 'https://tolking.github.io' })
 
-fs.unlinkSync(toAbsolute('dist/static/ssr-manifest.json'))
+  for (const path of files) {
+    const url = path
+      .replace(/^docs\/docs/, '')
+      .replace(/\.(vue|md)$/, '')
+      .replace(/index$/, '')
+      .replace(/\/([^/]*)$/, (item) =>
+        item.replace(/\B([A-Z])/g, '-$1').toLowerCase()
+      )
+    const filePath = `dist/static${url.replace(/\/$/, '/index')}.html`
+    const lang = url.match(/^\/([\w|-]*)\//)
+    const locale = lang ? lang[1] : 'en'
+    const [appHtml, preloadLinks] = await render(url, manifest)
+    const html = template
+      .replace('en', locale)
+      .replace('<title></title>\n', '')
+      .replace('<!--preload-links-->', preloadLinks)
+      .replace('<!--app-html-->', appHtml)
+
+    writeFileRecursive(toAbsolute(filePath), html)
+    stream.write({
+      url: 'element-pro-components' + url,
+      lastmod: await getUpdatedTime(path),
+      changefreq: 'monthly',
+      priority: 0.8,
+    })
+  }
+
+  stream.end()
+  streamToPromise(stream).then((data) => {
+    writeFileRecursive(toAbsolute('dist/static/sitemap.xml'), data.toString())
+  })
+  fs.unlinkSync(toAbsolute('dist/static/ssr-manifest.json'))
+})()