+++ docs/src/main.ts
@@ -1,8 +1,9 @@
-import Layout from './App.vue'
 import { App, createApp as _createApp, createSSRApp } from 'vue'
 import { createRouter } from './router/index'
-import { Router } from 'vue-router'
+import { createHead, HeadClient } from '@vueuse/head'
 import {
+  ElConfigProvider,
+  ElIcon,
   ElButton,
   ElInput,
   ElUpload,
@@ -12,24 +13,48 @@ import {
   ElDropdownMenu,
   ElDatePicker,
 } from 'element-plus'
+import {
+  Edit,
+  House,
+  TakeawayBox,
+  CaretTop,
+  CaretBottom,
+} from '@element-plus/icons'
 import ElementPro from '/@src/index'
 import ProCode from './components/ProCode.vue'
-import 'element-plus/lib/theme-chalk/index.css'
+import Layout from './App.vue'
+import 'element-plus/lib/theme-chalk/el-var.css'
+import 'element-plus/lib/theme-chalk/el-icon.css'
+import 'element-plus/lib/theme-chalk/el-button.css'
+import 'element-plus/lib/theme-chalk/el-input.css'
+import 'element-plus/lib/theme-chalk/el-upload.css'
+import 'element-plus/lib/theme-chalk/el-tag.css'
+import 'element-plus/lib/theme-chalk/el-dropdown.css'
+import 'element-plus/lib/theme-chalk/el-dropdown-item.css'
+import 'element-plus/lib/theme-chalk/el-dropdown-menu.css'
+import 'element-plus/lib/theme-chalk/el-date-picker.css'
+import '/@src/styles/index.css'
 import './styles/index.css'
+import type { Router } from 'vue-router'
 
 export function createApp(): {
   app: App<Element>
   router: Router
+  head: HeadClient
 } {
   const app =
     import.meta.env.MODE === 'production'
       ? createSSRApp(Layout)
       : _createApp(Layout)
   const router = createRouter()
+  const head = createHead()
 
   app
     .use(router)
+    .use(head)
     .component('ProCode', ProCode)
+    .use(ElConfigProvider)
+    .use(ElIcon)
     .use(ElButton)
     .use(ElInput)
     .use(ElUpload)
@@ -38,7 +63,12 @@ export function createApp(): {
     .use(ElDropdownItem)
     .use(ElDropdownMenu)
     .use(ElDatePicker)
+    .component('el-icon-edit', Edit)
+    .component('el-icon-house', House)
+    .component('el-icon-takeaway-box', TakeawayBox)
+    .component('el-icon-caret-top', CaretTop)
+    .component('el-icon-caret-bottom', CaretBottom)
     .use(ElementPro)
 
-  return { app, router }
+  return { app, router, head }
 }