+++ docs/src/views/Form.vue
@@ -34,7 +34,9 @@
 </template>
 
 <script setup lang="ts">
-import { h, onMounted, ref } from 'vue'
+import { h, onMounted, ref, shallowRef } from 'vue'
+import { ElSwitch } from 'element-plus'
+import 'element-plus/lib/theme-chalk/el-switch.css'
 import type {
   IFormColumns,
   IFormMenuColumns,
@@ -71,7 +73,54 @@ const list = [
   { value: 'JavaScript', tag: 'javascript' },
   { value: 'Python', tag: 'python' },
 ]
-const columns = ref<IFormColumns>([
+const treeList = [
+  {
+    label: '1',
+    value: 1,
+    disabled: true,
+    children: [
+      {
+        value: 11,
+        label: '1-1',
+        children: [{ value: '111', label: '1-1-1' }],
+      },
+    ],
+  },
+  {
+    value: 2,
+    label: '2',
+    children: [
+      {
+        value: 21,
+        label: '2-1',
+        disabled: true,
+        children: [{ value: 211, label: '2-1-1' }],
+      },
+      {
+        value: 22,
+        label: '2-2',
+        children: [{ value: 221, label: '2-2-1', disabled: true }],
+      },
+    ],
+  },
+  {
+    value: 3,
+    label: '3',
+    children: [
+      {
+        value: 31,
+        label: '3-1',
+        children: [{ value: 311, label: '3-1-1' }],
+      },
+      {
+        value: 32,
+        label: '3-2',
+        children: [{ value: 321, label: '3-2-1', disabled: true }],
+      },
+    ],
+  },
+]
+const columns = shallowRef<IFormColumns>([
   {
     // label: 'Slot',
     prop: 'slot',
@@ -207,6 +256,22 @@ const columns = ref<IFormColumns>([
       data: list,
       config: { label: 'tag' },
     },
+  },
+  {
+    label: 'pro-tree-select',
+    prop: 'treeSelect',
+    component: 'pro-tree-select',
+    span: 12,
+    props: {
+      data: treeList,
+      clearable: true,
+    },
+  },
+  {
+    label: 'el-switch',
+    prop: 'switch',
+    component: ElSwitch,
+    span: 12,
   },
 ])
 const columns1 = ref<IFormColumns<RuleForm>>([