+++ src/Breadcrumb/Breadcrumb.vue
@@ -4,22 +4,27 @@
       v-for="item in list"
       :key="item.path"
     >
-      <router-link
+      <pro-link
         v-if="item.meta"
         :to="item.path"
       >
         {{ item.meta.title }}
-      </router-link>
+      </pro-link>
     </el-breadcrumb-item>
   </el-breadcrumb>
 </template>
 
-<script setup name="ProBreadcrumb" lang="ts">
-import { computed, defineProps } from 'vue'
+<script lang="ts">
+export default { name: 'ProBreadcrumb' }
+</script>
+
+<script setup lang="ts">
+import { computed } from 'vue'
 import { useRoute } from 'vue-router'
 import { ElBreadcrumb, ElBreadcrumbItem } from 'element-plus'
 import { findRouterItemListByPath } from '../utils/index'
 import { useCurrentRoutes } from '../composables/index'
+import ProLink from '../Link/index'
 import type { IRouteRecordRaw } from '../types/index'
 
 const props = defineProps<{ routes?: IRouteRecordRaw[] }>()