+++ src/Layout/Layout.vue
@@ -25,18 +25,17 @@
       </template>
     </pro-layout-aside>
     <section class="pro-container">
-      <pro-layout-header @toggle-collapse="toggleShow">
+      <pro-layout-header
+        :collapse="show"
+        @toggle-collapse="toggleShow"
+      >
         <template #left>
-          <!-- TODO: remove slots `left-header` `right-header` `bottom-header` -->
-          <slot name="left-header" />
           <slot name="header-left" />
         </template>
         <template #right>
-          <slot name="right-header" />
           <slot name="header-right" />
         </template>
       </pro-layout-header>
-      <slot name="bottom-header" />
       <slot name="header-bottom" />
       <pro-layout-main :transition="transition">
         <template #top>
@@ -51,8 +50,12 @@
   </section>
 </template>
 
-<script setup name="ProLayout" lang="ts">
-import { defineProps, toRefs, useContext } from 'vue'
+<script lang="ts">
+export default { name: 'ProLayout' }
+</script>
+
+<script setup lang="ts">
+import { toRefs, useSlots } from 'vue'
 import ProLayoutAside from './LayoutAside.vue'
 import ProLayoutHeader from './LayoutHeader.vue'
 import ProLayoutMain from './LayoutMain.vue'
@@ -63,25 +66,7 @@ const props = defineProps<{
   transition?: string
 }>()
 const { collapse, transition } = toRefs(props)
-const { slots } = useContext()
+const slots = useSlots()
 const attrs = useAttrs()
 const { show, toggleShow } = useShow(collapse?.value)
 </script>
-
-<style lang="postcss">
-.pro-layout {
-  display: flex;
-  height: var(--layout-height);
-  overflow: hidden;
-  & .pro-container {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    width: calc(100% - var(--aside-width));
-    background: var(--c-page-background);
-  }
-  & .aside-collapse + .pro-container {
-    width: calc(100% - var(--aside-collapse-width));
-  }
-}
-</style>