+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from './composables/index'
 export * from './utils/index'
 export * from './types/index'
-import './styles/index.css'
 import type { App } from 'vue'
 import type { IDefineComponent, InstallOptions } from './types/index'
 
@@ -14,14 +13,16 @@ import ProCrud from './Crud/index'
 import ProForm from './Form/index'
 import ProInputTag from './InputTag/index'
 import ProLayout from './Layout/index'
+import ProLink from './Link/index'
 import ProMenu from './Menu/index'
 import ProRadio from './Radio/index'
 import ProRadioButton from './RadioButton/index'
 import ProSelect from './Select/index'
 import ProTable from './Table/index'
 import ProTabs from './Tabs/index'
+import ProTreeSelect from './TreeSelect/index'
 
-const components: Record<string, IDefineComponent> = {
+const components: IDefineComponent[] = [
   ProAutocompleteTag,
   ProBreadcrumb,
   ProCheckbox,
@@ -31,23 +32,25 @@ const components: Record<string, IDefineComponent> = {
   ProForm,
   ProInputTag,
   ProLayout,
+  ProLink,
   ProMenu,
   ProRadio,
   ProRadioButton,
   ProSelect,
   ProTable,
   ProTabs,
-}
+  ProTreeSelect,
+]
 
 const install = (app: App, options?: InstallOptions): void => {
   options && (app.config.globalProperties.$PROOPTIONS = options)
 
-  for (const key in components) {
-    const item = components[key]
-    app.component(item.name || key, item)
-  }
+  components.forEach((item) => {
+    app.component(item.name, item)
+  })
 }
 
+export default { install }
 export {
   ProAutocompleteTag,
   ProBreadcrumb,
@@ -58,13 +61,13 @@ export {
   ProForm,
   ProInputTag,
   ProLayout,
+  ProLink,
   ProMenu,
   ProRadio,
   ProRadioButton,
   ProSelect,
   ProTable,
   ProTabs,
+  ProTreeSelect,
   install,
 }
-
-export default { install }